+++ OndselSolver/ASMTCompoundJoint.h
@@ -18,7 +18,7 @@ namespace MbD {
         void parseASMT(std::vector<std::string>& lines) override;
         void readDistanceIJ(std::vector<std::string>& lines);
         void createMbD(std::shared_ptr<System> mbdSys, std::shared_ptr<Units> mbdUnits) override;
-        void storeOnLevel(std::ofstream& os, int level) override;
+        void storeOnLevel(std::ofstream& os, size_t level) override;
 
         double distanceIJ = 0.0;
