+++ OndselSolver/ASMTConstantGravity.cpp
@@ -44,7 +44,7 @@ void MbD::ASMTConstantGravity::setg(double a, double b, double c)
 	g = std::make_shared<FullColumn<double>>(ListD{ a, b, c });
 }
 
-void MbD::ASMTConstantGravity::storeOnLevel(std::ofstream& os, int level)
+void MbD::ASMTConstantGravity::storeOnLevel(std::ofstream& os, size_t level)
 {
 	storeOnLevelString(os, level, "ConstantGravity");
 	storeOnLevelArray(os, level + 1, *g);
