+++ OndselSolver/ASMTPart.cpp
@@ -118,7 +118,7 @@ void MbD::ASMTPart::createMbD(std::shared_ptr<System> mbdSys, std::shared_ptr<Un
 	if (isFixed) std::static_pointer_cast<Part>(mbdObject)->asFixed();
 }
 
-void MbD::ASMTPart::storeOnLevel(std::ofstream& os, int level)
+void MbD::ASMTPart::storeOnLevel(std::ofstream& os, size_t level)
 {
 	storeOnLevelString(os, level, "Part");
 	storeOnLevelName(os, level + 1);
@@ -133,7 +133,7 @@ void MbD::ASMTPart::storeOnLevel(std::ofstream& os, int level)
 	storeOnLevelRefSurfaces(os, level + 1);
 }
 
-void MbD::ASMTPart::storeOnLevelMassMarker(std::ofstream& os, int level)
+void MbD::ASMTPart::storeOnLevelMassMarker(std::ofstream& os, size_t level)
 {
 	principalMassMarker->storeOnLevel(os, level);
 }
