+++ OndselSolver/ASMTPart.h
@@ -25,8 +25,8 @@ namespace MbD {
         FColDsptr omeOpO() override;
         ASMTPart* part() override;
         void createMbD(std::shared_ptr<System> mbdSys, std::shared_ptr<Units> mbdUnits) override;
-        void storeOnLevel(std::ofstream& os, int level) override;
-        void storeOnLevelMassMarker(std::ofstream& os, int level);
+        void storeOnLevel(std::ofstream& os, size_t level) override;
+        void storeOnLevelMassMarker(std::ofstream& os, size_t level);
         void storeOnTimeSeries(std::ofstream& os) override;
 
         //std::shared_ptr<std::vector<std::shared_ptr<ASMTFeature>>> featureOrder;
