+++ OndselSolver/ASMTRefSurface.cpp
@@ -16,7 +16,7 @@ void MbD::ASMTRefSurface::parseASMT(std::vector<std::string>&)
 	assert(false);
 }
 
-void MbD::ASMTRefSurface::storeOnLevel(std::ofstream&, int)
+void MbD::ASMTRefSurface::storeOnLevel(std::ofstream&, size_t)
 {
 	assert(false);
 }
