+++ OndselSolver/ASMTScrewJoint.h
@@ -19,7 +19,7 @@ namespace MbD {
         void parseASMT(std::vector<std::string>& lines) override;
         void readPitch(std::vector<std::string>& lines);
         void createMbD(std::shared_ptr<System> mbdSys, std::shared_ptr<Units> mbdUnits) override;
-        void storeOnLevel(std::ofstream& os, int level) override;
+        void storeOnLevel(std::ofstream& os, size_t level) override;
 
         double pitch = 0.0, aConstant = 0.0;
     };
